+++ benches/parallel.rs
@@ -3,15 +3,13 @@
 extern crate cgmath;
 extern crate rand;
 extern crate specs;
-
 extern crate test;
 
 use cgmath::Vector2;
 use rand::thread_rng;
-use specs::{Component, DenseVecStorage, DispatcherBuilder, Entities, Entity, Fetch,
-            HashMapStorage, Join, NullStorage, ReadStorage, RunningTime, System, VecStorage,
-            World, WriteStorage};
-
+use specs::{Component, DenseVecStorage, DispatcherBuilder, Entities, Entity,
+            Fetch, HashMapStorage, Join, NullStorage, ReadStorage, RunningTime,
+            System, VecStorage, World, WriteStorage};
 use test::Bencher;
 
 type Vec2 = Vector2<f32>;