+++ src/storage/tests.rs
@@ -178,6 +178,26 @@ mod test {
         type Storage = BTreeStorage<Self>;
     }
 
+    #[derive(PartialEq, Eq, Debug)]
+    #[cfg(feature="rudy")]
+    struct CRudy(u32);
+    #[cfg(feature="rudy")]
+    impl From<u32> for CRudy {
+        fn from(v: u32) -> CRudy {
+            CRudy(v)
+        }
+    }
+    #[cfg(feature="rudy")]
+    impl AsMut<u32> for CRudy {
+        fn as_mut(&mut self) -> &mut u32 {
+            &mut self.0
+        }
+    }
+    #[cfg(feature="rudy")]
+    impl Component for CRudy {
+        type Storage = RudyStorage<Self>;
+    }
+
     #[derive(Clone, Debug)]
     struct Cnull(u32);
     impl Default for Cnull {
@@ -233,8 +253,8 @@ mod test {
 
         for i in 0..1_000 {
             *s.get_mut(Entity::new(i, Generation::new(1)))
-                 .unwrap()
-                 .as_mut() -= 718;
+                .unwrap()
+                .as_mut() -= 718;
         }
 
         for i in 0..1_000 {
@@ -398,6 +418,37 @@ mod test {
         test_clear::<CBtree>();
     }
 
+    #[cfg(feature="rudy")]
+    #[test]
+    fn rudy_test_add() {
+        test_add::<CRudy>();
+    }
+    #[cfg(feature="rudy")]
+    #[test]
+    fn rudy_test_sub() {
+        test_sub::<CRudy>();
+    }
+    #[cfg(feature="rudy")]
+    #[test]
+    fn rudy_test_get_mut() {
+        test_get_mut::<CRudy>();
+    }
+    #[cfg(feature="rudy")]
+    #[test]
+    fn rudy_test_add_gen() {
+        test_add_gen::<CRudy>();
+    }
+    #[cfg(feature="rudy")]
+    #[test]
+    fn rudy_test_sub_gen() {
+        test_sub_gen::<CRudy>();
+    }
+    #[cfg(feature="rudy")]
+    #[test]
+    fn rudy_test_clear() {
+        test_clear::<CRudy>();
+    }
+
     #[test]
     fn dummy_test_clear() {
         test_clear::<Cnull>();
@@ -497,7 +548,7 @@ mod test {
     }
 }
 
-#[cfg(feature="serialize")]
+#[cfg(feature = "serialize")]
 mod serialize_test {
     extern crate serde_json;