+++ CHANGELOG.md
@@ -1,4 +1,12 @@
 ## 0.9.3
+* Add `specs-derive` crate, custom `#[derive]` for components ([#192])
+* Add lazy updates: insert and remove components, execute closures on world ([#214], [#221])
+
+[#192]: https://github.com/slide-rs/specs/pull/192
+[#214]: https://github.com/slide-rs/specs/pull/214
+[#221]: https://github.com/slide-rs/specs/pull/221
+
+## 0.9.2
 * Fixed grammar in book ([#198])
 * Better docs for `World` and better panic message ([#199])
 * Add support for Emscripten ([#205])